# server/openapi.yaml
openapi: 3.0.3
info:
  title: NetTechPro API
  description: |
    RESTful API cho website bán thiết bị mạng NetTechPro.
    
    ## Tính năng
    - Quản lý sản phẩm (Products)
    - Quản lý danh mục (Categories)
    - Quản lý thương hiệu (Brands)
    - Quản lý đơn hàng (Orders)
    - Giỏ hàng (Cart)
    - Xác thực (Authentication)
    
    ## Mock Data
    - 8 sản phẩm thiết bị mạng
    - 8 danh mục
    - 10 thương hiệu
    - 3 đơn hàng mẫu
  version: 1.0.0
  contact:
    name: NetTechPro API Support
    email: support@nettechpro.me
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.nettechpro.me
    description: Production server
...
paths:
...
  /api/products:
    get:
      summary: Lấy danh sách sản phẩm
      description: Lấy tất cả sản phẩm với filter, sort, search
      tags: [Products]
      parameters:
        - name: categoryId
          in: query
          description: Lọc theo danh mục
          schema:
            type: string
            example: cat-1
        - name: brandId
          in: query
          description: Lọc theo thương hiệu
          schema:
            type: string
            example: brand-1
        - name: featured
          in: query
          description: Lọc sản phẩm nổi bật
          schema:
            type: boolean
        - name: search
          in: query
          description: Tìm kiếm theo tên
          schema:
            type: string
        - name: sortBy
          in: query
          description: Sắp xếp
          schema:
            type: string
            enum: [price-asc, price-desc, name, newest]
      responses:
        '200':
          description: Danh sách sản phẩm
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  total:
                    type: integer
                    example: 8
    post:
      summary: Tạo sản phẩm mới
      description: Tạo sản phẩm (Admin only)
      tags: [Products]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '201':
          description: Sản phẩm đã được tạo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
  /api/products/{id}:
    get:
      summary: Lấy chi tiết sản phẩm
      tags: [Products]
      parameters:
        - name: id
          in: path
          required: true
          description: ID hoặc slug của sản phẩm
          schema:
            type: string
      responses:
        '200':
          description: Chi tiết sản phẩm
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Không tìm thấy sản phẩm
    put:
      summary: Cập nhật sản phẩm
      tags: [Products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '200':
          description: Sản phẩm đã được cập nhật
    delete:
      summary: Xóa sản phẩm
      tags: [Products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Sản phẩm đã được xóa
...
components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: string
          example: prod-1
        name:
          type: string
          example: Cisco Catalyst 2960-X Series Switch
        nameVi:
          type: string
          example: Switch Cisco Catalyst 2960-X 24 Port
        slug:
          type: string
          example: cisco-catalyst-2960-x-24-port
        description:
          type: string
        descriptionVi:
          type: string
        price:
          type: number
          example: 15900000
        salePrice:
          type: number
          example: 14500000
        discount:
          type: number
          example: 9
        stock:
          type: integer
          example: 25
        categoryId:
          type: string
        brandId:
          type: string
        images:
          type: array
          items:
            type: string
        rating:
          type: number
          example: 4.8
        featured:
          type: boolean
        new:
          type: boolean
        bestseller:
          type: boolean
    ProductInput:
      type: object
      required: [name, nameVi, price, stock, categoryId, brandId]
      properties:
        name:
          type: string
        nameVi:
          type: string
        slug:
          type: string
        description:
          type: string
        descriptionVi:
          type: string
        price:
          type: number
        salePrice:
          type: number
        stock:
          type: integer
        categoryId:
          type: string
        brandId:
          type: string
        images:
          type: array
          items:
            type: string
        featured:
          type: boolean
        new:
          type: boolean
        bestseller:
          type: boolean
...